test: Add cherrypick function

+++ tests/utils/test_repo.py
@@ -1,3 +1,4 @@
+"""Wrapper around the git.Repo class that implements concrete methods for unit testing."""
 import logging
 import os
 import shutil
@@ -6,7 +7,7 @@ from pathlib import Path
 from tempfile import mkdtemp
 from uuid import uuid4
 
-from git import Repo
+from git import Commit, Repo
 
 logger = logging.getLogger('wemogy.get-release-version-action.tests.repo')
 
@@ -28,9 +29,11 @@ class TestRepo:
     """Wrapper around the git.Repo class that implements concrete methods for unit testing."""
     path: Path
     repo: Repo
+    keep_repository_dir: bool
 
-    def __init__(self) -> None:
+    def __init__(self, keep_repository_dir: bool = False) -> None:
         """Wrapper around the git.Repo class that implements concrete methods for unit testing."""
+        self.keep_repository_dir = keep_repository_dir
         self.path = Path(mkdtemp(prefix='wemogy.get-release-version-action.tests'))
         logger.info('Created git repository in directory %s', self.path)
         os.chdir(self.path)
@@ -52,12 +55,17 @@ class TestRepo:
         self.close()
 
     def close(self) -> None:
-        """Remove the test repo"""
+        """Remove the test repo."""
         self.repo.close()
-        shutil.rmtree(self.path)
+
+        if not self.keep_repository_dir:
+            logger.info('Removing repository %s', self.path)
+            shutil.rmtree(self.path)
+        else:
+            logger.info('Keeping repository %s', self.path)
 
     def create_branch(self, name: str, base_name: str | None) -> None:
-        """Create a branch named `name` that is based on the branch with `base_name` and check it out"""
+        """Create a branch named `name` that is based on the branch with `base_name` and check it out."""
         if base_name is not None:
             self.checkout(base_name)
         self.repo.create_head(name)
@@ -67,39 +75,49 @@ class TestRepo:
         """
         Checkout the branch with the specified name.
 
-        :raises GitBranchNotFoundError: If the branch was not found
+        :raises GitBranchNotFoundError: If the branch was not found.
         """
         try:
-            branch = self.repo.heads[branch_name]
+            self.repo.heads[branch_name].checkout()
         except IndexError as exc:
             raise GitBranchNotFoundError(f'Branch {branch_name} was not found') from exc
 
-        self.repo.head.reference = branch
-
-    def commit(self, commit_message: CommitMessages | str, file_name: str | None = None) -> None:
+    def commit(self, commit_message: CommitMessages | str, file_name: str | None = None) -> Commit:
         """
         Create a file and make a commit with the given level as commit message.
 
-        :param commit_message: A conventional commit message
-        :param file_name: An optional file name, if None a random file name is chosen
+        :param commit_message: A conventional commit message.
+        :param file_name: An optional file name, if None a random file name is chosen.
+        :returns: The created commit.
         """
         file_name = file_name or f'file_{uuid4()}'
         file_path = self.path / file_name
         file_path.write_text('test')
 
         self.repo.index.add(file_name)
-        self.repo.index.commit(commit_message)
+        return self.repo.index.commit(commit_message)
 
     def merge(self, source_branch_name: str, dest_branch_name: str) -> None:
         """
-        Merge a branch into another branch.
-        This also checks out the destination branch.
+        Merge a branch into another branch and check out the destination branch.
 
         :param source_branch_name: The branch name to merge from.
         :param dest_branch_name: The branch name to merge into.
+        :raises GitBranchNotFoundError: If the branch was not found.
         """
         self.checkout(dest_branch_name)
         self.repo.git.merge(source_branch_name)
+
+    def cherrypick(self, commit: Commit, dest_branch_name: str) -> None:
+        """
+        Cherrypick a commit into a branch and check out the destination branch.
+
+        :param commit: The commit to cherrypick.
+        :param dest_branch_name: The branch name to cherrypick onto.
+        :raises GitBranchNotFoundError: If the branch was not found.
+        """
+        self.checkout(dest_branch_name)
+        self.repo.git.cherry_pick(commit.hexsha)
 
 
 __all__ = [
